--- backend/api/components/components.go
@@ -1,1 +1,0 @@
-package components
