+++ backend/api/config/api.ini
@@ -1,0 +1,2 @@
+[API]
+rest_api_enabled = true
