--- backend/api/config/config.go
@@ -1,1 +1,0 @@
-package config
