--- backend/api/handler/routes.go
@@ -1,1 +1,0 @@
-package routes
